Fixed major in rule
To make sure there was made no attempt to apply rules to atomic
formulas, the atomic formulas were filtered out. I forgot that these of
course have to be put back into the sequent's, or else that information
is lost. This made the function a lot less clean (code-wise), so I hope
to refactor this.